requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
PyMuPDF==1.21.1
pandas==1.3.3
openpyxl==3.0.9
Pillow==8.3.2
pytesseract==0.3.8 